refactor(handler): move run DTO mappers to runs_dto.go (#1022)
Pure code motion: the run/message DTO mapping cluster (isPlanningPhase,
runPriorityClass, runToDTO, failureSnapshotWebURL, messageToDTO,
steerInputToDTO) moves verbatim from workers.go into the new runs_dto.go.
The now-unused forgesvc import is dropped from workers.go (its only use was
runToDTO). --color-moved clean; no exported-API change and no behaviour
change (go doc -all byte-identical, tests unchanged). No positional
comments needed fixing: the only above/below references inside the moved
blocks are intra-function field-mapping notes within runToDTO's body, which
move with it.
